# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Firetower",
    "description": "The Firetower control plane: API, scheduling, and worker transports.",
    "contact": {
      "name": "Westlabs LLC"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/providers/{id}": {
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Sign out: forget the token and stop any authorization still waiting.",
        "operationId": "disconnect_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/authorize": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Start an authorization and begin waiting for it to be approved.",
        "description": "Returns immediately with the code to show. The waiting happens here rather\nthan in the browser so that closing the tab doesn't abandon it.",
        "operationId": "authorize_provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingAuth"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/client-id": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Register an application to authorize against.",
        "description": "Asked twice on purpose: once in the setup wizard, where it is skippable,\nand again on the connect-a-repository screen at the moment somebody wants\nthe thing it enables. Stored rather than configured, so it takes effect\nwithout a restart.",
        "operationId": "set_client_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientId"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Stored"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{id}/repos": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "What the authorized account can see. This is the picker's data.",
        "operationId": "list_provider_repos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoteRepo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Every non-success response, so failures are as typed as everything else.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "For humans and logs. The interface should switch on `code` and write its\nown copy — only it knows the context and what to offer next."
          }
        }
      },
...
      "ClientId": {
        "type": "object",
        "description": "The client id for a git host, supplied by whoever is setting this up.",
        "required": [
          "clientId"
        ],
        "properties": {
          "clientId": {
            "type": "string",
            "description": "GitHub's looks like `Ov23li…`. Public by design: a device-flow\napplication has no paired secret."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "The catalogue is the type, so there is no separate list to keep in sync.",
        "enum": [
          "InvalidRequest",
          "NotFound",
          "NoCapacity",
          "HostUnreachable",
          "RepoNotConnected",
          "SessionEnded",
          "ProviderNotConfigured",
          "ProviderNotConnected",
          "RepoAccessDenied",
          "RepoUnreachable",
          "RepoUnusable",
          "RepoInUse",
          "ActionFailed",
          "Unauthorized",
          "PasswordChangeRequired",
          "Internal"
        ]
      },
...
      "PendingAuth": {
        "type": "object",
        "description": "A device authorization waiting for someone to approve it in a browser.",
        "required": [
          "userCode",
          "verificationUri"
        ],
        "properties": {
          "userCode": {
            "type": "string",
            "description": "The short code to type. Shown, not clicked."
          },
          "verificationUri": {
            "type": "string",
            "description": "Where to type it."
          }
        }
      },
...
      "RemoteRepo": {
        "type": "object",
        "description": "A repository the authorized account can see.",
        "required": [
          "slug",
          "remote",
          "defaultBranch",
          "private"
        ],
        "properties": {
          "defaultBranch": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "pushedAt": {
            "type": [
              "string",
              "null"
            ],
            "description": "Most recently pushed first is the order people actually want."
          },
          "remote": {
            "type": "string",
            "description": "The https URL, since a token authenticates over https."
          },
          "slug": {
            "type": "string",
            "description": "`acme/backend`"
          }
        }
      },
...
    }
  }
}